refactor: consolidate hot path implementations
- Remove unused hot path implementations (hot_path.rs, hot_path_streaming.rs, streaming.rs)
- Rename streaming_simple.rs to hot_path.rs as the single implementation
- Update module structure and imports
- This consolidation reduces code duplication and confusion

The streaming_simple implementation was the most complete and production-ready,
already being used by the ProxyService.
